Hotfix/CHO-62 (#90)
* Refactor components and add category buttons

* Increment current index and question count, display dialog
