+++ src/components/header.tsx
@@ -1,0 +1,42 @@
+import React from "react";
+import { Button } from "@/components/ui/button";
+import Logo from "@/components/logo";
+import { ModeToggle } from "@/components/mode-toggle";
+import { LogIn } from "lucide-react";
+import { useRouter } from "next/navigation";
+import { useUser } from "@auth0/nextjs-auth0/client";
+
+export default function Header() {
+    const { user } = useUser();
+    const router = useRouter();
+
+    return (
+        <header className="flex flex-row items-center justify-between">
+            <Logo />
+            <div className="flex flex-row items-center gap-4">
+                <ModeToggle />
+                {!user ? (
+                    <Button
+                        className="gap-2"
+                        onClick={() => {
+                            router.push("/api/auth/login");
+                        }}
+                    >
+                        <LogIn className="h-4 w-4" />
+                        Log In
+                    </Button>
+                ) : (
+                    <Button
+                        className="gap-2"
+                        onClick={() => {
+                            router.push("/api/auth/logout");
+                        }}
+                    >
+                        <LogIn className="h-4 w-4" />
+                        Log out
+                    </Button>
+                )}
+            </div>
+        </header>
+    );
+}
